requirements-comfy.txt
accelerate==0.27.0
gradio==3.41.2
huggingface-hub==0.20.2
numpy==1.24.4
omegaconf==2.3.0
safetensors==0.3.3
transformers==4.33.1
easydict==1.13
ftfy==6.2.0
pillow
